Fixes #160 - Changes check in appcontypes to same originator
Previous in the appcontypes, while getting a connection, it was checked
if a timed out connection had the same connection triad. This was wrong
as a timed out connection is not existing anymore, and each connection
has a unique connection serial number. In order to enable that the
original originator and only the original originator can reclaim the
timed out connection, first the check has been changes to equal
originator (vendor ID and serial number) and in the case that it is not
the original originator an ownership conflict error is returned. With
this fix the original originator can now reclaim a exclusive connection,
as well as new originator are not able to claim the exclusive owner
connection, without a reset or the original originator deletes his
connection.

Signed-off-by: Martin Melik-Merkumians <melik-merkumians@acin.tuwien.ac.at>