Corrects doxygen entries and moves semicolon to OPENER_ASSERT macro
When OPENER_TRACES are deactivated, the static code checking tools will
report false positive empty line errors, as the OPENER_ASSERT will be
removed, but the semicolon will stay. With this change the semicolon for
OPENER_ASSERT is now part of the macro and is removed from the code,
resulting in complete empty lines, if OPENER_TRACES is deactivated.

Signed-off-by: Martin Melik-Merkumians <melik-merkumians@acin.tuwien.ac.at>