Changes strtok to strtok_r
The networkconfig made use of the now dicouraged strtok function, which
is non-reentrant. This fix replaces it with the reentrant verion
strtok_r. In addition the way how files are read and processed has been
changed from a line-wise process to reading the whole file upfront and
processing it afterwards, simplifying the processing part.

Signed-off-by: Martin Melik-Merkumians <melik-merkumians@acin.tuwien.ac.at>