Correct the call signature of executeEventLoop()
This patch corrects the call signature for the executeEventLoop() function
to match the call signature that is required for the start_routine()
argument. Also its Doxygen documentation comment is corrected.

executeEventLoop() now returns a valid pointer to the static
pthread_dummy_return variable to avoid returning a random pointer as it
happened before.

The real time thread priority of the stack thread is reduced from 80 to 25.
This is for Linux running with the RT / lowlatency configuration enabled
where the interrupt backend threads run with a priority of 50.

Signed-off-by: Stefan Mätje <stefan.maetje@esd.eu>