Remove dead code from GetAttributeSingleTcpIpInterface()
The mentioned function is specific for the TCP/IP object and is therefore
always called with instances of a TCP/IP object.

That means the if block will never be entered because it tests for having
an instance of the Assembly class. This seems to be a copy'n'paste
leftover of the special handling done in GetAttributeSingle() of
cipcommon.c for the Assembly class's Data attribute.

Signed-off-by: Stefan Mätje <stefan.maetje@esd.eu>