Fix multicast IO connections on devices with multiple network interfaces
Without specifying a route for multicast packets outgoing multicast packets
are sent via the network interface that has the default route set. On a
machine with multiple interfaces this may not be the interface the
EtherNet/IP network is connected to.

To send the outgoing multicast packets on the right interface the socket
is configured to send the packets on the interface that has our own
EtherNet/IP IP address attached using IP_MULTICAST_IF as optname.

Signed-off-by: Stefan Mätje <stefan.maetje@esd.eu>