Avoid TRACE warning messages triggered by wrong use of CloseSocket()
In this POSIX specific source CloseSocket() was uses to close plain not
connected sockets. That use triggered warning messages like: "Could not
close socket 3 - Error Code: 107 - Transport endpoint is not connected"

This POSIX only code now uses a plain close() which seems to be right
here.

Also in ConfigureNetworkInterface() added a block in the source to let
"fd" go out of scope immediately after the close().

Signed-off-by: Stefan Mätje <stefan.maetje@esd.eu>