Revert "Issue#185: Memory double-free in multi-instance classes"
This reverts commit 88038c4739ea69ecde655973b9f0fb036a0d9448.

This fixed the memory double-free for multi-instance classes but only for
the case if the multiple instances were allocated at once when
CreateCipClass() is called with "number_of_instances" > 0 and no more
instances are added later.

Calling CreateAssemblyClass() multiple times like the samplelapplication.c
does creates multiple single instance blocks that are not freed later
except the first one.

The right fix would be NOT to allocate the needed CipInstances en bloc and
keep the code here as it was.

Signed-off-by: Stefan Mätje <stefan.maetje@esd.eu>