Fix: Second part to avoid memory management problems with multi-instance classes
This patch is the second part of the fix and works together with the commit
reverted as commit 5a95bffa50485cbb46fe10e779a84c3acdea238a.

This patch now allocates in AddCipInstances() the instances for a class one
by one instead of allocating a block with multiple instances. This matches
the code to free the instances in DeleteAllClasses().

Before this patch AddCipInstances() allocated a block of 1-n instances at
once for each of its calls. But there was no information present anywhere
which of the instances had to be freed being the first in the block and
which not.

Signed-off-by: Stefan Mätje <stefan.maetje@esd.eu>