Cleanup: Fix warnings about wrong types of pointer parameters
Fix warnings in encap.c by adding the needed type casts. The C interface
of many socket functions has a "struct sockaddr *" pointer parameter
because "struct sockaddr" is intended as "super-type" for all the socket
address structures used. Therefore always this cast is needed and desired
for this type of function calls.

Signed-off-by: Stefan Mätje <stefan.maetje@esd.eu>