Fix the HandleReceivedExplict[Tcp|Udp]Data() return type inconsistency
The mentioned functions sometimes returned only values of EipStatus type
but sometimes the length of the response message to be send. But the
size of the response is in any case already transferred to the caller by
the outgoing_message.used_message_length structure member.

Now both functions return only values of EipStatus type and the callers
evaluate only this kind of status. To make the functions return EipStatus
consistently some lower level functions also had to be changed to return
EipStatus notably NotifyConnectedCommonPacketFormat() and
NotifyCommonPacketFormat().

Some changes are marked with TODO and therefore have to be discussed before
the final commit.

Signed-off-by: Stefan Mätje <stefan.maetje@esd.eu>