Fix: QoS DSCP values should come into effect only after a restart
The QoS DSCP values can be changed from the EIP network but the changes
should come into effect only after a restart. Therefore we take a copy
here when the network is initialized in NetworkHandlerInitialize() and
CipQosGetDscpPriority() is called the first time.

This flaw affected only the CIP IO connections because the other stuff uses
only fixed sockets with DSCP EXPLICIT set in NetworkHandlerInitialize().

Changed the function name to CipQosGetDscpPriority() to get a CipQos name
prefix.

Signed-off-by: Stefan Mätje <stefan.maetje@esd.eu>