Revert "Cleanup: Object's class ids get same type and naming convention"
This reverts commit 29a15cf61c94d48784badc89b235594551484284.
The revert is done to avoid conflicts with the changes coming from the
master branch when the branch Origin/Issue259 will be committed to the
master.

After the git rebase it showed up that revert had not been necessary
because the Issue259 branch already reverted these changes. So this
commit could have been empty if I would have not made some small
changes in comments and whitespace. Therefore amended this message.

Signed-off-by: Stefan Mätje <stefan.maetje@esd.eu>