Change the QoS module to provide access to a global QoS object instance
The QoS module was changed to provide access to a global QoS object
to allow the Non Volatile Data code to access the needed data.

In NetworkHandlerInitialize() we call CipQosUpdateUsedSetQosValues() to
activate the current DSCP values for use with CipQosGetDscpPriority().

Between the call of CipStackInit() and NetworkHandlerInitialize() the
NV data shall be loaded.

Some internal renaming to make the names a little shorter was done also.

Signed-off-by: Stefan Mätje <stefan.maetje@esd.eu>