Fix: ConfigureDomainName() did not work at all
Fixed a memory leak on read error of resolv.conf. The function did not
evaluate resolv.conf correctly and depended on the order of the occurrence
of keywords.

Now the function decodes "nameserver" lines and "domain" lines correctly.
Additionally the domain name is set from the first domain mentioned on a
line with the "search" keyword.

Signed-off-by: Stefan Mätje <stefan.maetje@esd.eu>