Fix: ConfigureNetworkInterface() did not find gateway
The read line from /proc/net/route was not split correctly into tokens. The
gateway value could never be read correctly. Also because the string to
struct in_addr conversion was tried with inet_pton() which is wrong. The
gateway string is a simple hex number in network byte order.

Changed function to do these two things the right way. Also a open failure
for /proc/net/route was not handled.

Signed-off-by: Stefan Mätje <stefan.maetje@esd.eu>