Change Ethernet Link object's handling of capabilities and interface label
The interface label is now part of the global g_ethernet_link object data.

Also a meta structure CipEthernetLinkMetaInterfaceCapability for the interface capabilities is added to the Ethernet Link object structure. This
meta attribute uses a bit mask to mark which entries of type CipEthernetLinkSpeedDuplexArrayEntry of the "speed_duplex_table" array are
valid and should be returned for a GetAttribute request to the interface
capability attribute (#11).

Signed-off-by: Stefan Mätje <stefan.maetje@esd.eu>