Add OPENER_IS_DLR_DEVICE as configuration option
OpENer can now be configured with OPENER_IS_DLR_DEVICE that should change
OpENer to become a DLR capable device. This is still work in progress.

At the moment it only adds needed defines in the opener_user_conf.h files
and the Ethernet Link object gains the capability to have an interface
label that is not empty.

Signed-off-by: Stefan Mätje <stefan.maetje@esd.eu>