Added Ethernet Link object's counter attributes and GetAndClear service
This commit adds the Interface Counters attribute (#4) and the Media
Counters attribute (#5) to the implementation of the Ethernet Link object.
This feature is enabled with OPENER_ETHLINK_CNTRS_ENABLE that is
automatically enabled if OpENer is configured with OPENER_IS_DLR_DEVICE
from the CMake command line.

At the moment only a dummy implementation to read the counters is provided
because the way to retrieve the real counters is hardware and platform
dependent. The dummy implementation reports bogus counter values that make
the validation of the data transport easy.

Signed-off-by: Stefan Mätje <stefan.maetje@esd.eu>