Add the implementation and initialization of the DLR object
This patch implements the DLR object for a non supervisor and non gateway
device. The implemented revision of the DLR device is 3. Therefore the
optional attributes 17, 18 and 19 needed only for revision 4 are
NOT implemented.

The DLR object is activated and used from cipcommon.c if the define
OPENER_IS_DLR_DEVICE has been set with -DOPENER_IS_DLR_DEVICE:BOOL=ON
from the CMake command line.

Implemented Attributes
----------------------
- Attribute  1: Network Topology
- Attribute  2: Network Status
- Attribute 10: Active Supervisor Address
- Attribute 12: Capability Flags

Implemented Services
--------------------
- GetAttributesAll
- GetAttributeSingle

Signed-off-by: Stefan Mätje <stefan.maetje@esd.eu>