Fix: GetConnectionId() returned only the lower 16 bit of the connection id
The function GetConnectionId() shall create a unique (even over reboots)
Connection Id of 32 bit length. This function uses g_incarnation_id as
proposed in Vol. 2 Section 3-3.7.1.2 "Using an Incarnation ID" to create
a 32 bit Connection Id.

The failure of the implementation was only to return the lower 16 bit of
the created value! Therefore after a reboot even if the g_incarnation_id
changed the changed part of the Connection Id was stripped from the
result on return from the function.

Now that my device can do a reboot the Conformance Test of the Connection
Manager failed.

Signed-off-by: Stefan Mätje <stefan.maetje@esd.eu>