Fix: Rejected Listen Only Connection even if another IO Connection exists
OpENer rejected to open a Listen Only Connection even if another IO
connection existed because it looked only for Multicast connections and
ignored existing Point-To-Point connections.

Looking at the table 3-6.2 in Vol. 1 Section 3-6.3 it can be derived that
the establishment of a new Listen Only connection should succeed for both
cases (existing connection in Multicast or Point-to-Point mode).

Renamed the function GetExistingProducerMulticastConnection() to
GetExistingProducerIoConnection() and provided a parameter that controls
if it should look for Multicast connections only or return also
Point-To-Point connections. Adjusted the use of the function accordingly.

Signed-off-by: Stefan Mätje <stefan.maetje@esd.eu>