Add support for Ethernet Link's Interface Control attribute needed for DLR
The spec requires the Interface Control attribute (#6) to be settable from
the EtherNet/IP interface because a DLR capable device needs to be able to
support forced 100MBit Full-Duplex connection.

Therefore added now the support to make this attribute settable and that it
saves its setting to the NV data.

Signed-off-by: Stefan Mätje <stefan.maetje@esd.eu>