Cleanup: Unify the use of the OPENER_ASSERT() macro and fixed some invalid uses
The standard C assert() macro is defined without a trailing semicolon. Some
of the OPENER_ASSERT() macro definitions had a trailing semicolon and some
not.

To make all variants of the macro definition interchangeable the definitions
in all opener_user_conf.h have been heavily reworked.

Now all OPENER_ASSERT() macro definitions lack a trailing semicolon. As a
consequence a semicolon had to be added on numerous uses of that macro in
the source.

At some points OPENER_ASSERT() was called with a literal character string
as argument. This failed the intended use completely and was replaced by
an OPENER_TRACE_ERR() call with the literal string followed by an
OPENER_ASSERT(false).

Signed-off-by: Stefan Mätje <stefan.maetje@esd.eu>