Fix: Return kCipErrorAttributeNotSupported for unsupported Ethernet Link object attributes
The SetAttributeSingleEthernetLink() function did only return kCipErrorAttributeNotSupported
for attributes that don't exist in the Ethernet Link object instance. But
there are some attributes that exist only to support the GetAttributesAll
service. For these attributes we also need to return kCipErrorAttributeNotSupported.
Therefore we return that error code if an attribute has not the
kGetableSingle property set.

Signed-off-by: Stefan Mätje <stefan.maetje@esd.eu>