Fix: Work around possible race condition when detecting gateway on POSIX
The function GetGatewayFromRoute() evaluates /proc/net/route to detect the
gateway. If the interface is configured by DHCP the DHCP client first sets
the IP address and then runs the scripts to setup the routing information.

The main program waits for the IP address to appear on the network interface
and immediately calls IfaceGetConfiguration() to get all needed information.
But it appears to be possible that GetGatewayFromRoute() is called before
the script setting up the routing information finishes.

This workaround adds a 300ms delay before calling GetGatewayFromRoute() to
fix that issue.

Signed-off-by: Manuel Köppen <manuel.koeppen@esd.eu>
Signed-off-by: Stefan Mätje <stefan.maetje@esd.eu>