Fix possible inconsistency between TCP/IP object attributes 3 and 5
The "Interface Configuration" attribute (#5) could be written even if the
"Configuration Control" attribute (#3) is not set for the "Static IP"
configuration method (its value is zero).

If such a SetAttributeSingle request for attribute #5 arrives the device
has two choices:

a) Reject the SetAttributeSingle request with the status codes 0x0C "Object
   state conflict" or 0x10 "Device state conflict".
b) Accept the SetAttributeSingle request but change the "Configuration
   Control" attribute (#3) to contain the configuration method value 0,
   i. e. "Static IP", on the fly.

We adopt the choice b) here.

Signed-off-by: Manuel Köppen <manuel.koeppen@esd.eu>
Signed-off-by: Stefan Mätje <stefan.maetje@esd.eu>