Fixes to pass the Off-Link Routing Test Procedure of Conformance Test
The OpENer stack is configured by default for a multicast TTL of 1 (TCP/IP
attribute #8) which is immutable atm.

A device that has a multicast TTL of 1 MUST check for a ForwardOpen request
that tries to establish a multicast T->O connection if the originator is on
the same subnet as the device itself. If this is not the case the sent multi-
cast packets will never reach the originator device.

Added this check in the ForwardOpen() function of cipconnectionmanager.c.
Because meanwhile the TCP/IP "Interface Configuration" is settable from the
CIP network the IP address the setting from g_tcpip.interface_configuration
can't be used for that test. It may have been changed between the call of
NetworkHandlerInitialize() and the ForwardOpen request.

Therefore save the IP address and network mask being valid during the
NetworkHandlerInitialize() call in the now extended g_network_status
structure.

The ListIdentity response also needs to include the "saved" IP address.
Changed the code in encap.c accordingly.

Signed-off-by: Manuel Köppen <manuel.koeppen@esd.eu>
Signed-off-by: Stefan Mätje <stefan.maetje@esd.eu>