Remove unnecessary local data pointer variables.
These pointers generated compiler warnings about differing const
qualifiers, e.g. Visual Studio C4090, due to a mismatch in the level of
pointer indirection. In other words, the local variable was a pointer to
a pointer, which was assigned the value(not address of) a pointer,
resulting in a bug when dereferenced by memcpy.