Make class name string dynamically-allocated.
This is intended to resolve compiler warnings about the CreateCipClass()
name parameter, which is an immutable(const) character string. Directly
assigning it to the CipClass struct class_name member causes warnings
because the class_name member is not const. Changing it to const would
cause a different set of warnings when the meta-class name, which is
already dynamically-allocated, is freed, because CipFree does not take a
const pointer argument.

Another benefit is callers of CreateCipClass() don't need to ensure the
string pointed to by the name parameter stays in scope for the lifetime
of the created CipClass.