Clear destination strings before decoding new content.
This is intended to fix a bug when decoding a CipString or a
CipShortString under the following cases:

1. The target string is empty, i.e., the string structure member is NULL.
   The previous memcpy() would receive NULL as the destination pointer.

2. The target string contains content that is shorter than the string
   to be decoded. The string structure member points to a dynamically-
   allocated space that is too small for the new string, therefore the
   previous memcpy() will overrun the allocated area.

The new approach calls SetCipStringByData() to ensure any previous
content is freed, and sufficient memory is allocated for the new
string.

The local string and short_string pointers were also removed as they
were just unnecessary duplicates of the original data parameter,
which was also renamed to be more descriptive.