Rename OPENER_BUILD_SHARED_LIBS -> BUILD_SHARED_LIBS
The BUILD_SHARED_LIBS is actually a cmake global option for the
add_library() command.  It changes all libraries to be built as
SHARED, unless specifically set as STATIC.

For details, see the online manual:

https://cmake.org/cmake/help/latest/variable/BUILD_SHARED_LIBS.html

This patch also changes from cache value to option(), still defaulting
to disabled/off/false.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>