cipioconnection: add the AllocateSocketAddressInfoItem function
OpenProducingMulticastConnection can possibly be called with a
common_packet_format_data structure that contains both address_info_item
with a non-zero type_id. When this happens, a reasonable alternative
would be to pick the address_info_item that has a type_id already set
the desired final value. Currently, OpenProducingMulticastConnection
will default to the first address_info_item, regardless of its content.

This commit factors out this scheme present in
OpenProducingMulticastConnection and OpenConsumingPointToPointConnection
and puts it in a separate AllocateSocketAddressInfoItem helper function.

Signed-off-by: Charles Perry <charles.perry@savoirfairelinux.com>