cip: accept the first consumed sequence number unconditionally
Accept the first consumed sequence number without comparing it to 0.

ODVA CT18.1 will send sequence numbers that starts at 0x80000000, which
will not pass SEQ_GT32 when compared to 0.

This commit adds a boolean flag to the cip_connection_object that
indicates if the eip_level_sequence_count_consuming counter was
initialized with data from the client.

Signed-off-by: Charles Perry <charles.perry@savoirfairelinux.com>