Fix #393: transfer multicast I/O connections also on timeout
When a connection is properly closed we transfer the socket ownership,
i.e., an input-only scanner can continue to receive multicast data from
the adapter if a master scanner closes its connection.  In contrast,
when a master scanner times out we did not properly transfer the socket.

This patch factors out the socket transfer from CloseIoConnection() to a
separate function for use also in HandleIoConnectionTimeOut().

Some minor cleanup for readability has been mixed in to this patch.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>